Refactor attacker/damage history entries.
Now DoT and one-time damage entries are represented differently.
- DoT really needs to store duration and damageRate;
- One-time attacks only need to store the amount of damage.

In particular this removes some nasty code that encodes (gem.cpp) the
duration in the damage field, and decodes it again (progressionmanager.cpp).
Not only is this confusing, it's easy for the two to get out of sync.

I'm pretty sure a minus sign mismatch caused bug #3351; should be fixed.
New assertions document (and check!) that values have the right sign.