Fix active magic display.
Use a private variable for the lastIndex instead of a static since it
applies to the state of buffList, not global state.

Use signed comparison to handle sequence number wrap around.

Remove redundant assignments to csString().

Use a reference to csArray<ActiveSpell*> to avoid copying the array.

Don't send registrationTime since it wasn't sent before and isn't used
by the client. Besides, csTicks on the server mean nothing to the client.

Remember to initialize the for loop variable i = 0;